ProjectMaster/Django/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1